Upgrade asana api client
ID in asana objects doesn't work anymore in favor of GID (string).
Because of this I have to change the code to use string for foreign objects id on all integrations.
Those foreign objects id are used in api call to third parties so it should be fine after adjusting api calls on all third parties.